Add v2 api tests for get job with filters (#173)
* reenable v2 filter tests

* clarify docstring and test cases
