+++ qiskit/providers/ibmq/api_v2/ibmqclient.py
@@ -272,8 +272,8 @@ class IBMQClient:
             job_id (str): the id of the job.
             excluded_fields (list[str]): names of the fields to explicitly
                 exclude from the result.
-            included_fields (list[str]): names of the fields to explicitly
-                include in the result.
+            included_fields (list[str]): names of the fields, if present, to explicitly
+                include in the result. All the other fields will not be included in the result.
 
         Returns:
             dict: job information.
