+++ test/ibmq/test_ibmq_connector.py
@@ -145,75 +145,6 @@ class TestIBMQConnector(QiskitTestCase):
         version = api.api_version()
         self.assertIn('new_api', version)
 
-    @requires_qe_access
-    @requires_classic_api
-    def test_get_job_includes(self, qe_token, qe_url):
-        """Check the include fields parameter for get_job."""
-        IBMQ.enable_account(qe_token, qe_url)
-
-        api, job = self._submit_job_to_backend('ibmq_qasm_simulator')
-        job_id = job['id']
-
-        # Get the job, including a field.
-        self.assertIn('deleted', job)
-        self.assertIn('shots', job)
-        job_included = api.get_job(job_id, include_fields=['deleted'])
-        # Ensure the result has only the included field
-        self.assertIn('deleted', job_included)
-        self.assertNotIn('shots', job_included)
-
-    @requires_qe_access
-    @requires_classic_api
-    def test_get_job_excludes(self, qe_token, qe_url):
-        """Check the exclude fields parameter for get_job."""
-        IBMQ.enable_account(qe_token, qe_url)
-
-        api, job = self._submit_job_to_backend('ibmq_qasm_simulator')
-        job_id = job['id']
-
-        # Get the job, excluding a field.
-        self.assertIn('shots', job)
-        self.assertIn('deleted', job)
-        job_excluded = api.get_job(job_id, exclude_fields=['deleted'])
-        # Ensure the result only excludes the specified field
-        self.assertNotIn('deleted', job_excluded)
-        self.assertIn('shots', job)
-
-    @requires_qe_access
-    @requires_classic_api
-    def test_get_job_with_invalid_filter(self, qe_token, qe_url):
-        """Check get_job with an invalid filter."""
-        IBMQ.enable_account(qe_token, qe_url)
-
-        api, job = self._submit_job_to_backend('ibmq_qasm_simulator')
-        job_id = job['id']
-
-        # Get the job, excluding an non-existent field.
-        self.assertNotIn('dummy_exclude', job)
-        self.assertNotIn('dummy_include', job)
-        filtered_job = api.get_job(job_id, exclude_fields=['dummy_exclude'],
-                                   include_fields=['dummy_include'])
-        self.assertFalse(filtered_job)
-
-    def _submit_job_to_backend(self, backend_name):
-        """Submit a generic qobj job to the backend
-
-        Args:
-            backend_name (str): backend name
-
-        Returns:
-            tuple(IBMQConnector, dict):
-                IBMQConnector: API for communicating with IBMQ.
-                dict: API response to the job submit.
-        """
-        backend = IBMQ.get_backend(backend_name)
-        qobj = assemble(transpile([self.qc1, self.qc2], backend=backend, seed_transpiler=self.seed),
-                        backend=backend, shots=1)
-
-        api = backend._api
-        job = api.submit_job(qobj.to_dict(), backend_name)
-        return api, job
-
 
 class TestAuthentication(QiskitTestCase):
     """Tests for the authentication features.
